+++ ee/apps/account-service/src/Account.ts
@@ -1,10 +1,8 @@
-import { ServiceClass } from '@rocket.chat/core-services';
+import { ServiceClass, Settings } from '@rocket.chat/core-services';
 import type { IAccount, ILoginResult } from '@rocket.chat/core-services';
-import { Settings } from '@rocket.chat/models';
 import { getLoginExpirationInDays } from '@rocket.chat/tools';
 
 import { loginViaResume } from './lib/loginViaResume';
-import { loginViaUsername } from './lib/loginViaUsername';
 import { removeSession } from './lib/removeSession';
 
 export class Account extends ServiceClass implements IAccount {
@@ -22,13 +20,9 @@ export class Account extends ServiceClass implements IAccount {
 		});
 	}
 
-	async login({ resume, user, password }: { resume: string; user: { username: string }; password: string }): Promise<false | ILoginResult> {
+	async login({ resume }: { resume: string }): Promise<false | ILoginResult> {
 		if (resume) {
 			return loginViaResume(resume, this.loginExpiration);
-		}
-
-		if (user && password) {
-			return loginViaUsername(user, password, this.loginExpiration);
 		}
 
 		return false;
@@ -39,8 +33,8 @@ export class Account extends ServiceClass implements IAccount {
 	}
 
 	override async started(): Promise<void> {
-		const expiry = await Settings.findOne({ _id: 'Accounts_LoginExpiration' }, { projection: { value: 1 } });
+		const expiry = await Settings.get<number>('Accounts_LoginExpiration');
 
-		this.loginExpiration = getLoginExpirationInDays(expiry?.value as number);
+		this.loginExpiration = getLoginExpirationInDays(expiry);
 	}
 }