+++ ee/packages/omni-core-ee/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/omni-core-ee
 
+## 0.0.15
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-services@0.12.3
+  - @rocket.chat/models@2.0.3
+  - @rocket.chat/omni-core@0.0.15
+  </details>
+
 ## 0.0.14
 
 ### Patch Changes